+++ aioxmpp/security_layer.py
@@ -104,6 +104,7 @@ import OpenSSL.SSL
 from . import errors, sasl, stream_xsos, xso, protocol
 from .utils import namespaces
 
+
 logger = logging.getLogger(__name__)
 
 
@@ -666,28 +667,6 @@ class ErrorRecordingVerifier(CertificateVerifier):
                     ", ".join(map(str, self._errors))))
 
 
-@stream_xsos.StreamFeatures.as_feature_class
-class STARTTLSFeature(xso.XSO):
-    class STARTTLSRequired(xso.XSO):
-        TAG = (namespaces.starttls, "required")
-
-    TAG = (namespaces.starttls, "starttls")
-
-    required = xso.Child([STARTTLSRequired])
-
-
-class STARTTLS(xso.XSO):
-    TAG = (namespaces.starttls, "starttls")
-
-
-class STARTTLSFailure(xso.XSO):
-    TAG = (namespaces.starttls, "failure")
-
-
-class STARTTLSProceed(xso.XSO):
-    TAG = (namespaces.starttls, "proceed")
-
-
 class STARTTLSProvider:
     """
     A TLS provider to negotiate STARTTLS on an existing XML stream. This
@@ -738,7 +717,7 @@ class STARTTLSProvider:
         """
 
         try:
-            feature = features[STARTTLSFeature]
+            feature = features[stream_xsos.StartTLSFeature]
         except KeyError:
             return self._fail_if_required("STARTTLS not supported by peer")
 
@@ -751,11 +730,11 @@ class STARTTLSProvider:
         response = yield from protocol.send_and_wait_for(
             xmlstream,
             [
-                STARTTLS()
+                stream_xsos.StartTLS()
             ],
             [
-                STARTTLSFailure,
-                STARTTLSProceed,
+                stream_xsos.StartTLSFailure,
+                stream_xsos.StartTLSProceed,
             ]
         )